Add files via upload
This is a rough beginning for extracting events from a CinePlex csv file.  In a nutshell the program reads the data into a pandas data frame object, then timestamps marking entry and exit events of various targets are extracted using numerical filtering operations.  Events are to be used as reference points for extracting peri-event averages of concurrent calcium imaging signals.   The csv file contained in this repository is representative data from an example recording session.